Convert course modules to JSX
